transcrever/requirements.txt
Flask==2.0.1
python-docx==0.8.11
assemblyai==0.20.0
celery==5.3.4
redis==4.6.0
docx2pdf==0.1.8
gunicorn==20.1.0
python-dotenv==0.19.1
Werkzeug==2.0.1
flask-cors==3.0.10
pydantic==1.10.8